requirements.txt
packaging==23.1
Pillow==9.5.0
pip-review==1.3.0
pyocr==0.8.3
tesseract==0.1.3
